+++ src/index.ts
@@ -39,10 +39,9 @@ if(client.config.logs?.enabled) {
 
 client.on("ready", async () => {
     await connection.connect().then(async () => {
-        //console.log(await connection.query("drop table user_data"))
-
         await connection.query("CREATE TABLE IF NOT EXISTS user_data (index SERIAL, user_id VARCHAR(100) PRIMARY KEY, consent bool DEFAULT true, tokens int NOT NULL DEFAULT 0, cost double precision default 0, blacklisted bool DEFAULT false)")
         await connection.query("CREATE TABLE IF NOT EXISTS chats (index SERIAL, id VARCHAR(100) PRIMARY KEY, user_id VARCHAR(100) NOT NULL, messages JSON[] DEFAULT '{}', created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP)")
+        await connection.query("CREATE TABLE IF NOT EXISTS assistant_threads (index SERIAL, id text PRIMARY KEY, channel_id text NOT NULL, owner_id text NOT NULL)")
 
         console.log("Tables created")
     }).catch(console.error);
@@ -50,7 +49,7 @@ client.on("ready", async () => {
     client.commands.loadClasses().catch(console.error)
     client.components.loadClasses().catch(console.error)
     client.contexts.loadClasses().catch(console.error)
-    //client.modals.loadClasses().catch(console.error)
+    client.modals.loadClasses().catch(console.error)
     client.user?.setPresence({activities: [{type: ActivityType.Listening, name: "to ChatGPT screaming at your requests"}], status: PresenceUpdateStatus.DoNotDisturb })
     console.log(`Ready as ${client.user?.username}`)